report_generator/requirements.txt
Django==2.1.2
django-graphos==0.3.41
django-progressbarupload==0.1.7
et-xmlfile==1.0.1
jdcal==1.4
openpyxl==2.5.9
pytz==2018.7
